fix test: fix tests in SongFacadeTest. test: create SongControllerIntegrationTest
