+++ src/test/java/com/nique1/musicmanagementsystem/song/application/api/SongControllerTest.java
@@ -316,71 +316,6 @@ class SongControllerTest {
                 .andExpect(jsonPath("$").isEmpty());
     }
 
-    @Test
-    void shouldReturnSongDataGivenPartialLowercaseArtistName() throws Exception {
-        //given
-        given(songFacade.getSongByArtistTrackOrYear("ade", null, null, null))
-                .willReturn(List.of(new SongRspDto(uuid, "Adele", "Hello", 295, 2010)));
-        //when-then
-        mockMvc.perform(get("/songs")
-                        .queryParam("artist-name", "ade"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].uuid").value("d6e80790-081a-4abd-b5e7-1ac0cedc9ebd"))
-                .andExpect(jsonPath("$[0].artistName").value("Adele"))
-                .andExpect(jsonPath("$[0].trackName").value("Hello"))
-                .andExpect(jsonPath("$[0].trackLength").value(295))
-                .andExpect(jsonPath("$[0].year").value(2010));
-    }
-
-
-    @Test
-    void shouldReturnSongDataGivenPartialLowercaseTrackName() throws Exception {
-        //given
-        given(songFacade.getSongByArtistTrackOrYear(null, "hell", null, null))
-                .willReturn(List.of(new SongRspDto(uuid, "Adele", "Hello", 295, 2010)));
-        //when-then
-        mockMvc.perform(get("/songs")
-                        .queryParam("track-name", "hell"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].uuid").value("d6e80790-081a-4abd-b5e7-1ac0cedc9ebd"))
-                .andExpect(jsonPath("$[0].artistName").value("Adele"))
-                .andExpect(jsonPath("$[0].trackName").value("Hello"))
-                .andExpect(jsonPath("$[0].trackLength").value(295))
-                .andExpect(jsonPath("$[0].year").value(2010));
-    }
-
-    @Test
-    void shouldReturnSongDataGivenPartialLowercaseTrackNameAndYearFrom() throws Exception {
-        //given
-        given(songFacade.getSongByArtistTrackOrYear(null, "hell", 2010, null))
-                .willReturn(List.of(new SongRspDto(uuid, "Adele", "Hello", 295, 2010)));
-        //when-then
-        mockMvc.perform(get("/songs")
-                        .queryParam("track-name", "hell")
-                        .queryParam("year-from", "2010"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].uuid").value("d6e80790-081a-4abd-b5e7-1ac0cedc9ebd"))
-                .andExpect(jsonPath("$[0].artistName").value("Adele"))
-                .andExpect(jsonPath("$[0].trackName").value("Hello"))
-                .andExpect(jsonPath("$[0].trackLength").value(295))
-                .andExpect(jsonPath("$[0].year").value(2010));
-    }
-
-    @Test
-    void shouldReturnSongDataGivenPartialUpperCaseArtistName() throws Exception {
-        //given
-        given(songFacade.getSongByArtistTrackOrYear("ADE", null, null, null))
-                .willReturn(List.of(new SongRspDto(uuid, "Adele", "Hello", 295, 2010)));
-        //when-then
-        mockMvc.perform(get("/songs")
-                        .queryParam("artist-name", "ADE"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].uuid").value("d6e80790-081a-4abd-b5e7-1ac0cedc9ebd"))
-                .andExpect(jsonPath("$[0].artistName").value("Adele"))
-                .andExpect(jsonPath("$[0].trackName").value("Hello"))
-                .andExpect(jsonPath("$[0].trackLength").value(295))
-                .andExpect(jsonPath("$[0].year").value(2010));
-    }
 
 
 }
